requirements.txt
blinker==1.8.2
click==8.1.7
drf-spectacular==0.27.2
drf-yasg==1.21.8
itsdangerous==2.2.0
jinja2==3.1.4
openapi==2.0.0
pip-chill==1.0.3
pipdeptree==2.24.0
psycopg2-binary==2.9.10
werkzeug==2.3.8

Django~=5.1.4
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1